requirements.txt
python-telegram-bot==21.3
requests==2.32.3
redis==5.0.6
python-dotenv==1.0.1
